#216 : Deprecate customEntityInstanceService.
- Add deprecation notes.
- Refactor some codes to use crossStorageService.
