+++ meveo-admin/web/src/main/java/org/meveo/admin/action/admin/custom/CustomEntityInstanceBean.java
@@ -194,7 +194,7 @@ public class CustomEntityInstanceBean extends CustomFieldBean<CustomEntityInstan
 
 			customFieldDataEntryBean.saveCustomFieldsToEntity(entity, isNew);
 
-			boolean checkBeforeUpdate = customEntityInstanceService.checkBeforeUpdate(repository, entity);
+			boolean checkBeforeUpdate = crossStorageService.checkBeforeUpdate(repository, entity);
 			if (!checkBeforeUpdate) {
 				messages.error(new BundleKey("messages", "customEntityInstance.update.error"));
 				return null;
